64-bit: fix types to build for 64-bit targets
Some API within Audio HAL are inconsistant, switching from size_t
to uint32_t, preventing to switch to 64-bit.

This patch restores the consistency of the whole chain by switching
to size_t.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9545
Reviewed-on: https://android.intel.com:443/291420

+++ audio_conversion/include/AudioConversion.hpp
@@ -84,8 +84,8 @@ public:
      */
     android::status_t convert(const void *src,
                               void **dst,
-                              const uint32_t inFrames,
-                              uint32_t *outFrames);
+                              const size_t inFrames,
+                              size_t *outFrames);
 
     /**
      * Converts audio samples and output an exact number of output frames.
@@ -103,7 +103,7 @@ public:
      * @return status OK, error code otherwise.
      */
     android::status_t getConvertedBuffer(void *dst,
-                                         const uint32_t outFrames,
+                                         const size_t outFrames,
                                          android::AudioBufferProvider *bufferProvider);
 
 private:
