+++ audio_conversion/src/Resampler.cpp
@@ -144,8 +144,8 @@ void Resampler::convertFloat2Short(float *inp, int16_t *out, size_t sz) const
 
 status_t Resampler::resampleFrames(const void *src,
                                    void *dst,
-                                   const uint32_t inFrames,
-                                   uint32_t *outFrames)
+                                   const size_t inFrames,
+                                   size_t *outFrames)
 {
     size_t outFrameCount = convertSrcToDstInFrames(inFrames);
